+++ hydra-cluster/src/Hydra/Cluster/Scenarios.hs
@@ -108,8 +108,9 @@ import Hydra.Ledger.Cardano.Evaluate (maxTxExecutionUnits)
 import Hydra.Logging (Tracer, traceWith)
 import Hydra.Options (DirectChainConfig (..), startChainFrom)
 import Hydra.Tx (HeadId, IsTx (balance), Party, txId)
-import Hydra.Tx.ContestationPeriod (ContestationPeriod (UnsafeContestationPeriod), fromNominalDiffTime)
-import Hydra.Tx.DepositDeadline (DepositDeadline (..))
+import Hydra.Tx.ContestationPeriod (ContestationPeriod (UnsafeContestationPeriod))
+import Hydra.Tx.ContestationPeriod qualified as CP
+import Hydra.Tx.DepositPeriod (DepositPeriod (..))
 import Hydra.Tx.Utils (dummyValidatorScript, verificationKeyToOnChainId)
 import HydraNode (
   HydraClient (..),
@@ -179,17 +180,17 @@ oneOfThreeNodesStopsForAWhile tracer workDir cardanoNode hydraScriptsTxId = do
   seedFromFaucet_ cardanoNode carolCardanoVk 100_000_000 (contramap FromFaucet tracer)
 
   let contestationPeriod = UnsafeContestationPeriod 1
-  let depositDeadline = UnsafeDepositDeadline 10
+  let depositPeriod = DepositPeriod 10
   aliceChainConfig <-
-    chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [Bob, Carol] contestationPeriod depositDeadline
+    chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [Bob, Carol] contestationPeriod depositPeriod
       <&> setNetworkId networkId
 
   bobChainConfig <-
-    chainConfigFor Bob workDir nodeSocket hydraScriptsTxId [Alice, Carol] contestationPeriod depositDeadline
+    chainConfigFor Bob workDir nodeSocket hydraScriptsTxId [Alice, Carol] contestationPeriod depositPeriod
       <&> setNetworkId networkId
 
   carolChainConfig <-
-    chainConfigFor Carol workDir nodeSocket hydraScriptsTxId [Alice, Bob] contestationPeriod depositDeadline
+    chainConfigFor Carol workDir nodeSocket hydraScriptsTxId [Alice, Bob] contestationPeriod depositPeriod
       <&> setNetworkId networkId
 
   withHydraNode hydraTracer aliceChainConfig workDir 1 aliceSk [bobVk, carolVk] [1, 2, 3] $ \n1 -> do
@@ -253,13 +254,13 @@ restartedNodeCanObserveCommitTx tracer workDir cardanoNode hydraScriptsTxId = do
   seedFromFaucet_ cardanoNode bobCardanoVk 100_000_000 (contramap FromFaucet tracer)
 
   let contestationPeriod = UnsafeContestationPeriod 1
-  let depositDeadline = UnsafeDepositDeadline 200
+  let depositPeriod = DepositPeriod 200
   aliceChainConfig <-
-    chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [Bob] contestationPeriod depositDeadline
+    chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [Bob] contestationPeriod depositPeriod
       <&> setNetworkId networkId
 
   bobChainConfig <-
-    chainConfigFor Bob workDir nodeSocket hydraScriptsTxId [Alice] contestationPeriod depositDeadline
+    chainConfigFor Bob workDir nodeSocket hydraScriptsTxId [Alice] contestationPeriod depositPeriod
       <&> setNetworkId networkId
 
   let hydraTracer = contramap FromHydraNode tracer
@@ -285,9 +286,9 @@ restartedNodeCanAbort :: Tracer IO EndToEndLog -> FilePath -> RunningNode -> [Tx
 restartedNodeCanAbort tracer workDir cardanoNode hydraScriptsTxId = do
   refuelIfNeeded tracer cardanoNode Alice 100_000_000
   let contestationPeriod = UnsafeContestationPeriod 2
-  let depositDeadline = UnsafeDepositDeadline 200
+  let depositPeriod = DepositPeriod 200
   aliceChainConfig <-
-    chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [] contestationPeriod depositDeadline
+    chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [] contestationPeriod depositPeriod
       -- we delibelately do not start from a chain point here to highlight the
       -- need for persistence
       <&> modifyConfig (\config -> config{networkId, startChainFrom = Nothing})
@@ -316,13 +317,13 @@ nodeReObservesOnChainTxs tracer workDir cardanoNode hydraScriptsTxId = do
   tip <- queryTip networkId nodeSocket
   let contestationPeriod = UnsafeContestationPeriod 2
   let deadline = 10
-  let depositDeadline = UnsafeDepositDeadline deadline
+  let depositPeriod = DepositPeriod deadline
   aliceChainConfig <-
-    chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [Bob] contestationPeriod depositDeadline
+    chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [Bob] contestationPeriod depositPeriod
       <&> modifyConfig (\config -> config{networkId, startChainFrom = Nothing})
 
   bobChainConfig <-
-    chainConfigFor Bob workDir nodeSocket hydraScriptsTxId [Alice] contestationPeriod depositDeadline
+    chainConfigFor Bob workDir nodeSocket hydraScriptsTxId [Alice] contestationPeriod depositPeriod
       <&> modifyConfig (\config -> config{networkId, startChainFrom = Nothing})
 
   (aliceCardanoVk, aliceCardanoSk) <- keysFor Alice
@@ -395,7 +396,7 @@ nodeReObservesOnChainTxs tracer workDir cardanoNode hydraScriptsTxId = do
       pure (headId, decommitUTxO)
 
     bobChainConfigFromTip <-
-      chainConfigFor Bob workDir nodeSocket hydraScriptsTxId [Alice] contestationPeriod depositDeadline
+      chainConfigFor Bob workDir nodeSocket hydraScriptsTxId [Alice] contestationPeriod depositPeriod
         <&> modifyConfig (\config -> config{networkId, startChainFrom = Just tip})
 
     withTempDir "blank-state" $ \tmpDir -> do
@@ -452,10 +453,10 @@ singlePartyHeadFullLifeCycle tracer workDir node hydraScriptsTxId =
       refuelIfNeeded tracer node Alice 55_000_000
       -- Start hydra-node on chain tip
       tip <- queryTip networkId nodeSocket
-      contestationPeriod <- fromNominalDiffTime $ 10 * blockTime
-      let depositDeadline = UnsafeDepositDeadline 200
+      contestationPeriod <- CP.fromNominalDiffTime $ 10 * blockTime
+      let depositPeriod = DepositPeriod 200
       aliceChainConfig <-
-        chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [] contestationPeriod depositDeadline
+        chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [] contestationPeriod depositPeriod
           <&> modifyConfig (\config -> config{networkId, startChainFrom = Just tip})
       withHydraNode hydraTracer aliceChainConfig workDir 1 aliceSk [] [1] $ \n1 -> do
         -- Initialize & open head
@@ -510,9 +511,9 @@ singlePartyOpenAHead tracer workDir node hydraScriptsTxId callback =
     -- Start hydra-node on chain tip
     tip <- queryTip networkId nodeSocket
     let contestationPeriod = UnsafeContestationPeriod 100
-    let depositDeadline = UnsafeDepositDeadline 200
+    let depositPeriod = DepositPeriod 200
     aliceChainConfig <-
-      chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [] contestationPeriod depositDeadline
+      chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [] contestationPeriod depositPeriod
         <&> modifyConfig (\config -> config{networkId, startChainFrom = Just tip})
 
     (walletVk, walletSk) <- generate genKeyPair
@@ -551,8 +552,8 @@ singlePartyCommitsFromExternal tracer workDir node hydraScriptsTxId =
     $ do
       refuelIfNeeded tracer node Alice 25_000_000
       let contestationPeriod = UnsafeContestationPeriod 100
-      let depositDeadline = UnsafeDepositDeadline 200
-      aliceChainConfig <- chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [] contestationPeriod depositDeadline
+      let depositPeriod = DepositPeriod 200
+      aliceChainConfig <- chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [] contestationPeriod depositPeriod
       let hydraNodeId = 1
       let hydraTracer = contramap FromHydraNode tracer
       withHydraNode hydraTracer aliceChainConfig workDir hydraNodeId aliceSk [] [1] $ \n1 -> do
@@ -596,8 +597,8 @@ singlePartyUsesScriptOnL2 tracer workDir node hydraScriptsTxId =
     $ do
       refuelIfNeeded tracer node Alice 250_000_000
       let contestationPeriod = UnsafeContestationPeriod 1
-      let depositDeadline = UnsafeDepositDeadline 1
-      aliceChainConfig <- chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [] contestationPeriod depositDeadline
+      let depositPeriod = DepositPeriod 1
+      aliceChainConfig <- chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [] contestationPeriod depositPeriod
       let hydraNodeId = 1
       let hydraTracer = contramap FromHydraNode tracer
       withHydraNode hydraTracer aliceChainConfig workDir hydraNodeId aliceSk [] [1] $ \n1 -> do
@@ -716,8 +717,8 @@ singlePartyUsesWithdrawZeroTrick tracer workDir node hydraScriptsTxId =
       $ \utxoToCommit -> do
         -- Start hydra-node and open a head
         let contestationPeriod = UnsafeContestationPeriod 1
-        let depositDeadline = UnsafeDepositDeadline 1
-        aliceChainConfig <- chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [] contestationPeriod depositDeadline
+        let depositPeriod = DepositPeriod 1
+        aliceChainConfig <- chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [] contestationPeriod depositPeriod
         let hydraNodeId = 1
         let hydraTracer = contramap FromHydraNode tracer
         withHydraNode hydraTracer aliceChainConfig workDir hydraNodeId aliceSk [] [1] $ \n1 -> do
@@ -785,8 +786,8 @@ singlePartyCommitsScriptBlueprint tracer workDir node hydraScriptsTxId =
   (`finally` returnFundsToFaucet tracer node Alice) $ do
     refuelIfNeeded tracer node Alice 20_000_000
     let contestationPeriod = UnsafeContestationPeriod 100
-    let depositDeadline = UnsafeDepositDeadline 200
-    aliceChainConfig <- chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [] contestationPeriod depositDeadline
+    let depositPeriod = DepositPeriod 200
+    aliceChainConfig <- chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [] contestationPeriod depositPeriod
     let hydraNodeId = 1
     let hydraTracer = contramap FromHydraNode tracer
     (_, walletSk) <- keysFor AliceFunds
@@ -871,8 +872,8 @@ persistenceCanLoadWithEmptyCommit tracer workDir node hydraScriptsTxId =
   (`finally` returnFundsToFaucet tracer node Alice) $ do
     refuelIfNeeded tracer node Alice 20_000_000
     let contestationPeriod = UnsafeContestationPeriod 100
-    let depositDeadline = UnsafeDepositDeadline 200
-    aliceChainConfig <- chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [] contestationPeriod depositDeadline
+    let depositPeriod = DepositPeriod 200
+    aliceChainConfig <- chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [] contestationPeriod depositPeriod
     let hydraNodeId = 1
     let hydraTracer = contramap FromHydraNode tracer
     headId <- withHydraNode hydraTracer aliceChainConfig workDir hydraNodeId aliceSk [] [1] $ \n1 -> do
@@ -910,8 +911,8 @@ singlePartyCommitsFromExternalTxBlueprint tracer workDir node hydraScriptsTxId =
   (`finally` returnFundsToFaucet tracer node Alice) $ do
     refuelIfNeeded tracer node Alice 20_000_000
     let contestationPeriod = UnsafeContestationPeriod 100
-    let depositDeadline = UnsafeDepositDeadline 200
-    aliceChainConfig <- chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [] contestationPeriod depositDeadline
+    let depositPeriod = DepositPeriod 200
+    aliceChainConfig <- chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [] contestationPeriod depositPeriod
     let hydraNodeId = 1
     let hydraTracer = contramap FromHydraNode tracer
     (someExternalVk, someExternalSk) <- generate genKeyPair
@@ -972,9 +973,9 @@ canCloseWithLongContestationPeriod tracer workDir node hydraScriptsTxId = do
   -- Start hydra-node on chain tip
   tip <- queryTip networkId nodeSocket
   let oneWeek = UnsafeContestationPeriod (60 * 60 * 24 * 7)
-  let depositDeadline = UnsafeDepositDeadline 200
+  let depositPeriod = DepositPeriod 200
   aliceChainConfig <-
-    chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [] oneWeek depositDeadline
+    chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [] oneWeek depositPeriod
       <&> modifyConfig (\config -> config{networkId, startChainFrom = Just tip})
   let hydraTracer = contramap FromHydraNode tracer
   withHydraNode hydraTracer aliceChainConfig workDir 1 aliceSk [] [1] $ \n1 -> do
@@ -1009,8 +1010,8 @@ canSubmitTransactionThroughAPI tracer workDir node hydraScriptsTxId =
   (`finally` returnFundsToFaucet tracer node Alice) $ do
     refuelIfNeeded tracer node Alice 25_000_000
     let contestationPeriod = UnsafeContestationPeriod 100
-    let depositDeadline = UnsafeDepositDeadline 200
-    aliceChainConfig <- chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [] contestationPeriod depositDeadline
+    let depositPeriod = DepositPeriod 200
+    aliceChainConfig <- chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [] contestationPeriod depositPeriod
     let hydraNodeId = 1
     let hydraTracer = contramap FromHydraNode tracer
     withHydraNode hydraTracer aliceChainConfig workDir hydraNodeId aliceSk [] [hydraNodeId] $ \_ -> do
@@ -1064,9 +1065,9 @@ threeNodesNoErrorsOnOpen tracer tmpDir node@RunningNode{nodeSocket} hydraScripts
       hydraKeys = [aliceSk, bobSk, carolSk]
 
   let contestationPeriod = UnsafeContestationPeriod 2
-  let depositDeadline = UnsafeDepositDeadline 50
+  let depositPeriod = DepositPeriod 50
   let hydraTracer = contramap FromHydraNode tracer
-  withHydraCluster hydraTracer tmpDir nodeSocket 1 cardanoKeys hydraKeys hydraScriptsTxId contestationPeriod depositDeadline $ \clients -> do
+  withHydraCluster hydraTracer tmpDir nodeSocket 1 cardanoKeys hydraKeys hydraScriptsTxId contestationPeriod depositPeriod $ \clients -> do
     let leader = head clients
     waitForNodesConnected hydraTracer 20 clients
 
@@ -1105,9 +1106,9 @@ initWithWrongKeys workDir tracer node@RunningNode{nodeSocket} hydraScriptsTxId =
   (carolCardanoVk, _) <- keysFor Carol
 
   let contestationPeriod = UnsafeContestationPeriod 2
-  let depositDeadline = UnsafeDepositDeadline 200
-  aliceChainConfig <- chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [Carol] contestationPeriod depositDeadline
-  bobChainConfig <- chainConfigFor Bob workDir nodeSocket hydraScriptsTxId [Alice] contestationPeriod depositDeadline
+  let depositPeriod = DepositPeriod 200
+  aliceChainConfig <- chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [Carol] contestationPeriod depositPeriod
+  bobChainConfig <- chainConfigFor Bob workDir nodeSocket hydraScriptsTxId [Alice] contestationPeriod depositPeriod
 
   let hydraTracer = contramap FromHydraNode tracer
   withHydraNode hydraTracer aliceChainConfig workDir 3 aliceSk [bobVk] [3, 4] $ \n1 -> do
@@ -1140,12 +1141,12 @@ canCommit tracer workDir node hydraScriptsTxId =
       refuelIfNeeded tracer node Alice 30_000_000
       refuelIfNeeded tracer node Bob 30_000_000
       let contestationPeriod = UnsafeContestationPeriod 20
-      let depositDeadline = UnsafeDepositDeadline 200
+      let depositPeriod = DepositPeriod 200
       aliceChainConfig <-
-        chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [Bob] contestationPeriod depositDeadline
+        chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [Bob] contestationPeriod depositPeriod
           <&> setNetworkId networkId
       bobChainConfig <-
-        chainConfigFor Bob workDir nodeSocket hydraScriptsTxId [Alice] contestationPeriod depositDeadline
+        chainConfigFor Bob workDir nodeSocket hydraScriptsTxId [Alice] contestationPeriod depositPeriod
           <&> setNetworkId networkId
       withHydraNode hydraTracer aliceChainConfig workDir 1 aliceSk [bobVk] [2] $ \n1 -> do
         withHydraNode hydraTracer bobChainConfig workDir 2 bobSk [aliceVk] [1] $ \n2 -> do
@@ -1228,15 +1229,14 @@ canRecoverDeposit tracer workDir node hydraScriptsTxId =
     (`finally` returnFundsToFaucet tracer node Bob) $ do
       refuelIfNeeded tracer node Alice 30_000_000
       refuelIfNeeded tracer node Bob 30_000_000
-      -- NOTE: this value is also used to determine the deposit deadline
-      let deadline = 5
-      let contestationPeriod = UnsafeContestationPeriod deadline
-      let depositDeadline = UnsafeDepositDeadline 5
+      -- NOTE: Directly expire deposits
+      contestationPeriod <- CP.fromNominalDiffTime 1
+      let depositPeriod = DepositPeriod 1
       aliceChainConfig <-
-        chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [Bob] contestationPeriod depositDeadline
+        chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [Bob] contestationPeriod depositPeriod
           <&> setNetworkId networkId
       bobChainConfig <-
-        chainConfigFor Bob workDir nodeSocket hydraScriptsTxId [Alice] contestationPeriod depositDeadline
+        chainConfigFor Bob workDir nodeSocket hydraScriptsTxId [Alice] contestationPeriod depositPeriod
           <&> setNetworkId networkId
       withHydraNode hydraTracer aliceChainConfig workDir 1 aliceSk [bobVk] [2] $ \n1 -> do
         headId <- withHydraNode hydraTracer bobChainConfig workDir 2 bobSk [aliceVk] [1] $ \n2 -> do
@@ -1261,33 +1261,31 @@ canRecoverDeposit tracer workDir node hydraScriptsTxId =
         (balance <$> queryUTxOFor networkId nodeSocket QueryTip walletVk)
           `shouldReturn` lovelaceToValue commitAmount
 
-        resp <-
+        depositTransaction <-
           parseUrlThrow ("POST " <> hydraNodeBaseUrl n1 <> "/commit")
             <&> setRequestBodyJSON commitUTxO
               >>= httpJSON
-
-        let depositTransaction = getResponseBody resp :: Tx
+            <&> getResponseBody
 
         let tx = signTx walletSk depositTransaction
-
         submitTx node tx
 
-        waitForAllMatch 10 [n1] $ \v -> do
+        deadline <- waitForAllMatch 10 [n1] $ \v -> do
           guard $ v ^? key "tag" == Just "CommitRecorded"
+          v ^? key "deadline" >>= parseMaybe parseJSON
 
         (selectLovelace . balance <$> queryUTxOFor networkId nodeSocket QueryTip walletVk)
           `shouldReturn` 0
 
         let path = BSC.unpack $ urlEncode False $ encodeUtf8 $ T.pack $ show (getTxId $ getTxBody tx)
         -- NOTE: we need to wait for the deadline to pass before we can recover the deposit
-        -- NOTE: for some reason threadDelay on MacOS behaves differently than on Linux so we need + 1 here
-        threadDelay $ fromIntegral (deadline * 2 + 1)
+        diff <- realToFrac . diffUTCTime deadline <$> getCurrentTime
+        threadDelay $ diff + 1
 
-        recoverResp <-
+        (`shouldReturn` "OK") $
           parseUrlThrow ("DELETE " <> hydraNodeBaseUrl n1 <> "/commits/" <> path)
             >>= httpJSON
-
-        (getResponseBody recoverResp :: String) `shouldBe` "OK"
+            <&> getResponseBody @String
 
         waitForAllMatch 20 [n1] $ \v -> do
           guard $ v ^? key "tag" == Just "CommitRecovered"
@@ -1325,14 +1323,14 @@ canSeePendingDeposits tracer workDir node hydraScriptsTxId =
     (`finally` returnFundsToFaucet tracer node Bob) $ do
       refuelIfNeeded tracer node Alice 30_000_000
       refuelIfNeeded tracer node Bob 30_000_000
-      let deadline = 1
-      let contestationPeriod = UnsafeContestationPeriod deadline
-      let depositDeadline = UnsafeDepositDeadline 1
+      -- NOTE: Directly expire deposits
+      contestationPeriod <- CP.fromNominalDiffTime 1
+      let depositPeriod = DepositPeriod 1
       aliceChainConfig <-
-        chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [Bob] contestationPeriod depositDeadline
+        chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [Bob] contestationPeriod depositPeriod
           <&> setNetworkId networkId
       bobChainConfig <-
-        chainConfigFor Bob workDir nodeSocket hydraScriptsTxId [Alice] contestationPeriod depositDeadline
+        chainConfigFor Bob workDir nodeSocket hydraScriptsTxId [Alice] contestationPeriod depositPeriod
           <&> setNetworkId networkId
       withHydraNode hydraTracer aliceChainConfig workDir 1 aliceSk [bobVk] [2] $ \n1 -> do
         _ <- withHydraNode hydraTracer bobChainConfig workDir 2 bobSk [aliceVk] [1] $ \n2 -> do
@@ -1355,50 +1353,49 @@ canSeePendingDeposits tracer workDir node hydraScriptsTxId =
         commitUTxO2 <- seedFromFaucet node walletVk 4_000_000 (contramap FromFaucet tracer)
         commitUTxO3 <- seedFromFaucet node walletVk 3_000_000 (contramap FromFaucet tracer)
 
-        deposited <- flip execStateT [] $ forM [commitUTxO, commitUTxO2, commitUTxO3] $ \utxo -> do
-          resp <-
+        deposited <- forM [commitUTxO, commitUTxO2, commitUTxO3] $ \utxo -> do
+          depositTransaction <-
             parseUrlThrow ("POST " <> hydraNodeBaseUrl n1 <> "/commit")
               <&> setRequestBodyJSON utxo
                 >>= httpJSON
-
-          let depositTransaction = getResponseBody resp :: Tx
+              <&> getResponseBody
 
           let tx = signTx walletSk depositTransaction
+          let depositTxId = getTxId (getTxBody tx)
 
           liftIO $ submitTx node tx
 
           liftIO $ waitForAllMatch 10 [n1] $ \v ->
             guard $ v ^? key "tag" == Just "CommitRecorded"
 
-          pendingDepositReq <-
+          pendingDeposits <-
             parseUrlThrow ("GET " <> hydraNodeBaseUrl n1 <> "/commits")
               >>= httpJSON
+              <&> getResponseBody
 
-          let expectedResponse = getTxId (getTxBody tx)
-          _ <- modify (expectedResponse :)
-          expected <- get
-          let expectedResp = getResponseBody pendingDepositReq :: [TxId]
-          liftIO $ expectedResp `shouldBe` expected
+          liftIO $ pendingDeposits `shouldContain` [depositTxId]
+          pure depositTxId
 
         forM_ deposited $ \deposit -> do
           let path = BSC.unpack $ urlEncode False $ encodeUtf8 $ T.pack $ show deposit
+          -- XXX: should know the deadline from the query above
           -- NOTE: we need to wait for the deadline to pass before we can recover the deposit
-          threadDelay $ fromIntegral (deadline * 2)
-          recoverResp <-
+          threadDelay $ realToFrac (toNominalDiffTime depositPeriod * 2)
+
+          (`shouldReturn` "OK") $
             parseUrlThrow ("DELETE " <> hydraNodeBaseUrl n1 <> "/commits/" <> path)
               >>= httpJSON
-
-          (getResponseBody recoverResp :: String) `shouldBe` "OK"
+              <&> getResponseBody @String
 
           waitForAllMatch 10 [n1] $ \v -> do
             guard $ v ^? key "tag" == Just "CommitRecovered"
 
-        pendingDepositReq <-
+        pendingDeposits :: [TxId] <-
           parseUrlThrow ("GET " <> hydraNodeBaseUrl n1 <> "/commits")
             >>= httpJSON
+            <&> getResponseBody
 
-        let expectedResp = getResponseBody pendingDepositReq :: [TxId]
-        expectedResp `shouldBe` []
+        pendingDeposits `shouldBe` []
  where
   RunningNode{networkId, nodeSocket, blockTime} = node
 
@@ -1412,9 +1409,9 @@ canDecommit tracer workDir node hydraScriptsTxId =
   (`finally` returnFundsToFaucet tracer node Alice) $ do
     refuelIfNeeded tracer node Alice 30_000_000
     let contestationPeriod = UnsafeContestationPeriod 1
-    let depositDeadline = UnsafeDepositDeadline 200
+    let depositPeriod = DepositPeriod 200
     aliceChainConfig <-
-      chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [] contestationPeriod depositDeadline
+      chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [] contestationPeriod depositPeriod
         <&> setNetworkId networkId
     withHydraNode hydraTracer aliceChainConfig workDir 1 aliceSk [] [1] $ \n1 -> do
       -- Initialize & open head
@@ -1515,15 +1512,15 @@ canSideLoadSnapshot tracer workDir cardanoNode hydraScriptsTxId = do
   seedFromFaucet_ cardanoNode carolCardanoVk 100_000_000 (contramap FromFaucet tracer)
 
   let contestationPeriod = UnsafeContestationPeriod 1
-  let depositDeadline = UnsafeDepositDeadline 200
+  let depositPeriod = DepositPeriod 200
   aliceChainConfig <-
-    chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [Bob, Carol] contestationPeriod depositDeadline
+    chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [Bob, Carol] contestationPeriod depositPeriod
       <&> setNetworkId networkId
   bobChainConfig <-
-    chainConfigFor Bob workDir nodeSocket hydraScriptsTxId [Alice, Carol] contestationPeriod depositDeadline
+    chainConfigFor Bob workDir nodeSocket hydraScriptsTxId [Alice, Carol] contestationPeriod depositPeriod
       <&> setNetworkId networkId
   carolChainConfig <-
-    chainConfigFor Carol workDir nodeSocket hydraScriptsTxId [Alice, Bob] contestationPeriod depositDeadline
+    chainConfigFor Carol workDir nodeSocket hydraScriptsTxId [Alice, Bob] contestationPeriod depositPeriod
       <&> setNetworkId networkId
 
   withHydraNode hydraTracer aliceChainConfig workDir 1 aliceSk [bobVk, carolVk] [1, 2, 3] $ \n1 -> do
@@ -1640,13 +1637,13 @@ threeNodesWithMirrorParty tracer workDir cardanoNode@RunningNode{nodeSocket, net
   seedFromFaucet_ cardanoNode bobCardanoVk 100_000_000 (contramap FromFaucet tracer)
 
   let contestationPeriod = UnsafeContestationPeriod 1
-  let depositDeadline = UnsafeDepositDeadline 200
+  let depositPeriod = DepositPeriod 200
 
   aliceChainConfig <-
-    chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [Bob] contestationPeriod depositDeadline
+    chainConfigFor Alice workDir nodeSocket hydraScriptsTxId [Bob] contestationPeriod depositPeriod
       <&> setNetworkId networkId
   bobChainConfig <-
-    chainConfigFor Bob workDir nodeSocket hydraScriptsTxId [Alice] contestationPeriod depositDeadline
+    chainConfigFor Bob workDir nodeSocket hydraScriptsTxId [Alice] contestationPeriod depositPeriod
       <&> setNetworkId networkId
 
   let hydraTracer = contramap FromHydraNode tracer