+++ hydra-node/src/Hydra/API/ServerOutput.hs
@@ -25,6 +25,7 @@ import Hydra.Tx.ContestationPeriod (ContestationPeriod)
 import Hydra.Tx.Crypto (MultiSignature)
 import Hydra.Tx.IsTx (ArbitraryIsTx, IsTx (..))
 import Hydra.Tx.OnChainId (OnChainId)
+import Test.QuickCheck (recursivelyShrink)
 import Test.QuickCheck.Arbitrary.ADT (ToADTArbitrary)
 
 -- | The type of messages sent to clients by the 'Hydra.API.Server'.
@@ -181,10 +182,21 @@ data ServerOutput tx
   | DecommitInvalid {headId :: HeadId, decommitTx :: tx, decommitInvalidReason :: DecommitInvalidReason tx}
   | DecommitApproved {headId :: HeadId, decommitTxId :: TxIdType tx, utxoToDecommit :: UTxOType tx}
   | DecommitFinalized {headId :: HeadId, distributedUTxO :: UTxOType tx}
-  | CommitRecorded {headId :: HeadId, utxoToCommit :: UTxOType tx, pendingDeposit :: TxIdType tx, deadline :: UTCTime}
+  | -- XXX: Rename to DepositRecorded following the state events naming. But only
+    -- do this when changing the endpoint also to /commits
+    CommitRecorded
+      { headId :: HeadId
+      , utxoToCommit :: UTxOType tx
+      , -- XXX: Inconsinstent field name
+        pendingDeposit :: TxIdType tx
+      , deadline :: UTCTime
+      }
+  | DepositExpired {headId :: HeadId, depositTxId :: TxIdType tx, deadline :: UTCTime, chainTime :: UTCTime}
   | CommitApproved {headId :: HeadId, utxoToCommit :: UTxOType tx}
   | CommitFinalized {headId :: HeadId, depositTxId :: TxIdType tx}
-  | CommitRecovered {headId :: HeadId, recoveredUTxO :: UTxOType tx, recoveredTxId :: TxIdType tx}
+  | -- XXX: Rename to DepositRecovered to be more consistent. But only do this
+    -- when changing the endpoint also to /commits
+    CommitRecovered {headId :: HeadId, recoveredUTxO :: UTxOType tx, recoveredTxId :: TxIdType tx}
   | -- | Snapshot was side-loaded, and the included transactions can be considered final.
     -- The local state has been reset, meaning pending transactions were pruned.
     -- Any signing round has been discarded, and the snapshot leader has changed accordingly.
@@ -198,37 +210,7 @@ deriving anyclass instance IsChainState tx => ToJSON (ServerOutput tx)
 
 instance ArbitraryIsTx tx => Arbitrary (ServerOutput tx) where
   arbitrary = genericArbitrary
-
-  -- NOTE: Somehow, can't use 'genericShrink' here as GHC is complaining about
-  -- Overlapping instances with 'UTxOType tx' even though for a fixed `tx`, there
-  -- should be only one 'UTxOType tx'
-  shrink = \case
-    HeadIsInitializing headId xs -> HeadIsInitializing <$> shrink headId <*> shrink xs
-    Committed headId p u -> Committed <$> shrink headId <*> shrink p <*> shrink u
-    HeadIsOpen headId u -> HeadIsOpen <$> shrink headId <*> shrink u
-    HeadIsClosed headId s t -> HeadIsClosed <$> shrink headId <*> shrink s <*> shrink t
-    HeadIsContested headId sn dl -> HeadIsContested <$> shrink headId <*> shrink sn <*> shrink dl
-    ReadyToFanout headId -> ReadyToFanout <$> shrink headId
-    HeadIsFinalized headId u -> HeadIsFinalized <$> shrink headId <*> shrink u
-    HeadIsAborted headId u -> HeadIsAborted <$> shrink headId <*> shrink u
-    TxValid headId txid -> TxValid <$> shrink headId <*> shrink txid
-    TxInvalid headId u tx err -> TxInvalid <$> shrink headId <*> shrink u <*> shrink tx <*> shrink err
-    SnapshotConfirmed headId s ms -> SnapshotConfirmed <$> shrink headId <*> shrink s <*> shrink ms
-    IgnoredHeadInitializing{} -> []
-    DecommitRequested headId txid u -> DecommitRequested headId txid <$> shrink u
-    DecommitInvalid headId decommitTx decommitInvalidReason -> DecommitInvalid headId <$> shrink decommitTx <*> shrink decommitInvalidReason
-    DecommitApproved headId txid u -> DecommitApproved headId txid <$> shrink u
-    DecommitFinalized headId decommitTxId -> DecommitFinalized headId <$> shrink decommitTxId
-    CommitRecorded headId u i d -> CommitRecorded headId <$> shrink u <*> shrink i <*> shrink d
-    CommitApproved headId u -> CommitApproved headId <$> shrink u
-    CommitRecovered headId u rid -> CommitRecovered headId <$> shrink u <*> shrink rid
-    CommitFinalized headId depositTxId -> CommitFinalized headId <$> shrink depositTxId
-    NetworkConnected -> []
-    NetworkDisconnected -> []
-    NetworkVersionMismatch our theirs -> NetworkVersionMismatch <$> shrink our <*> shrink theirs
-    PeerConnected peer -> PeerConnected <$> shrink peer
-    PeerDisconnected peer -> PeerDisconnected <$> shrink peer
-    SnapshotSideLoaded headId snapshotNumber -> SnapshotSideLoaded headId <$> shrink snapshotNumber
+  shrink = recursivelyShrink
 
 instance (ArbitraryIsTx tx, IsChainState tx) => ToADTArbitrary (ServerOutput tx)
 
@@ -279,6 +261,7 @@ prepareServerOutput config response =
     DecommitFinalized{} -> encodedResponse
     DecommitInvalid{} -> encodedResponse
     CommitRecorded{} -> encodedResponse
+    DepositExpired{} -> encodedResponse
     CommitApproved{} -> encodedResponse
     CommitFinalized{} -> encodedResponse
     CommitRecovered{} -> encodedResponse