+++ hydra-node/src/Hydra/HeadLogic/Error.hs
@@ -16,6 +16,7 @@ import Hydra.Tx (
   SnapshotVersion,
  )
 import Hydra.Tx.Crypto (HydraKey, VerificationKey)
+import Test.QuickCheck (recursivelyShrink)
 
 data LogicError tx
   = UnhandledInput {input :: Input tx, currentHeadState :: HeadState tx}
@@ -34,7 +35,7 @@ instance
   Arbitrary (LogicError tx)
   where
   arbitrary = genericArbitrary
-  shrink = genericShrink
+  shrink = recursivelyShrink
 
 deriving stock instance
   ( Eq (HeadState tx)
@@ -70,7 +71,8 @@ data RequirementFailure tx
   | SnapshotAlreadySigned {knownSignatures :: [Party], receivedSignature :: Party}
   | AckSnNumberInvalid {requestedSn :: SnapshotNumber, lastSeenSn :: SnapshotNumber}
   | SnapshotDoesNotApply {requestedSn :: SnapshotNumber, txid :: TxIdType tx, error :: ValidationError}
-  | RecoverNotMatchingDeposit
+  | NoMatchingDeposit
+  | RequestedDepositNotActive {depositTxId :: TxIdType tx}
   deriving stock (Generic)
 
 deriving stock instance Eq (TxIdType tx) => Eq (RequirementFailure tx)