+++ CHANGELOG.md
@@ -8,6 +8,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 As a minor extension, we also keep a semantic version for the `UNRELEASED`
 changes.
 
+## [0.21.1] - UNRELEASED
+
+- **BREAKING** Fixed observation of deposit transactions:
+  - Correctly ignore deposits with deadlines in the past or too soon in the future.
+  - Replaced `--deposit-deadline` with `--deposit-period`.
+  - To make sure a deposit is picked up ensure that `--deposit-period` is longer than `--contestation-period`.
+  - Change persisted events of `hydra-node`.
+  - Change to the `ReqSn` message in the Hydra network protocol
+  - Added `DepositExpired` for when a deposit was deemed expired.
+
 ## [0.21.0] - 2025-04-28
 
 - **BREAKING** Switch to using `etcd` internally to establish a reliable L2 network